I once again forgot to commit during the process because I get into the zone.
I have added all the different leaders as images in the public folder.

I have added the basic layout for the dune page with all the leaders and created a redirect to the specific leader page if you select a leader.

Added in tailwind config why that cannot be used as a way to add background images dynamically.

Probably some other stuff too but I forgot what I have made :o)