Big commit...
Included a dark theme and a temp button for swapping the theme currently. Have chosen a new color theme following a color palette with a 12 step scale following the color theory from Radix.

I have started to move component to make use of Radix instead of other random icon libraries and or components.

I have converted the pomodoro page to fully include the new color theme and new components for popups. This is great because the new popups make use of Portals where as the ones created by me before did not and this will make things easier later on because of the added flexibility of portals.

I have deleted all the old files that were used for different tasks such as the old dropdowns as I'm now using Radix.

I have also started another section for the page. This will be for tracking the movies you've watched and giving them a rating. This might include shows as well and will be using the imdb api to get all the data.

I think I might start adding additional grid items just as the ideas popup for me as a way to keep track of them and then just leave them as NOT IMPLEMENTED on the frontpage so it's very apparent what I want to work on and I get that satisfaction of seeing it on the page immediatly. I think the pages might include my sketch up of the page unil they are implemented.

Start of moving to use React portals for popups.