Huge commit once again. A bit of a bad habbit of forgetting to commit as I'm just coding away from sun up to sun down basicly.
Created a Semester table in the database for storing all the databases.

Added a context to the app for the dialogs to use as a ref of the current element at the top so that the portals do not go above the dark mode class.

Created a UniversityGrid for all of the semesters at aau.

Created an image carousel for the semester demos.

Created the overall semester demo layout page.

ADded a few hooks for the semesters as well.