Fixed some more stuff to make it possible to publish this to the main branch since so much have changed. It is not completely done yet and not quite where I would like to have it before publishing it, but I can't send out resumes with my website and then not have the newest showing.
Some of the changes made seem to be a necessary evil since trpc and react query are acting weird. I'm not sure if that is an issue I've caused since I had to install on a machine that already had some old versions and something have been ruined because of that or what happened, but I will take a look at it at some later point.